move MetadataUrl to pywps.app.Common to avoid dependencies on sphinx
Steps to Reproduce
```bash
conda create -n pywps4.2.9 python=3.7
conda activate pywps4.2.9
python setup.py develop
python
```

Then
```python
from pywps import tests
Python 3.7.9 | packaged by conda-forge | (default, Dec  9 2020, 21:08:20)
[GCC 9.3.0] on linux
Type "help", "copyright", "credits" or "license" for more information.
>>> from pywps import tests
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/david/src/pywps/pywps/tests.py", line 15, in <module>
    from pywps.ext_autodoc import MetadataUrl
  File "/home/david/src/pywps/pywps/ext_autodoc.py", line 2, in <module>
    from sphinx.ext.autodoc import ClassDocumenter, bool_option
ModuleNotFoundError: No module named 'sphinx'
```
